Move PBXReferenceProxy to files package
